refresh token revocation

+++ src/server/src/main/kotlin/com/project/infrastructure/security/token/RefreshTokenResource.kt
@@ -3,6 +3,6 @@ package com.project.infrastructure.security.token
 data class RefreshTokenResource(
     val accessToken: String,
     val tokenType: String,
-    val expiresIn: Int,
+    val expiresIn: Long,
     val refreshToken: String,
 )
